+++ tests/test_tensor/test_contract.py
@@ -31,6 +31,113 @@ def test_tensor_contract_strip_exponent():
     assert m7 * 10**e7 == pytest.approx(z0)
 
 
+@pytest.mark.parametrize("strip_exponent", [False, True])
+@pytest.mark.parametrize("equalize_norms", [False, 1.0, True])
+@pytest.mark.parametrize("inplace", [False, True])
+def test_contract_tags_strip_exponent(
+    strip_exponent,
+    equalize_norms,
+    inplace,
+):
+    tn = qtn.TN_rand_reg(8, 3, 2)
+    Zex = tn.contract()
+
+    if inplace:
+        tnc = tn.copy()
+        tnc.contract_tags_(
+            all,
+            strip_exponent=strip_exponent,
+            equalize_norms=equalize_norms,
+        )
+        Z = tnc.arrays[0] * 10**tnc.exponent
+
+    else:
+        Z = tn.contract_tags(
+            all,
+            strip_exponent=strip_exponent,
+            equalize_norms=equalize_norms,
+        )
+        if strip_exponent:
+            Z = Z[0] * 10 ** Z[1]
+
+    assert Z == pytest.approx(Zex, rel=1e-3)
+
+
+@pytest.mark.parametrize("strip_exponent", [False, True])
+@pytest.mark.parametrize("equalize_norms", [False, 1.0, True])
+@pytest.mark.parametrize("inplace", [False, True])
+def test_contract_cumulative_strip_exponent(
+    strip_exponent,
+    equalize_norms,
+    inplace,
+):
+    mps = qtn.MPS_rand_state(7, 3)
+    tn = mps.make_norm()
+    Zex = tn.contract()
+
+    assert tn._CONTRACT_STRUCTURED
+
+    if inplace:
+        tnc = tn.copy()
+        tnc.contract_(
+            ...,
+            strip_exponent=strip_exponent,
+            equalize_norms=equalize_norms,
+        )
+        Z = tnc.arrays[0] * 10**tnc.exponent
+
+    else:
+        Z = tn.contract(
+            ...,
+            strip_exponent=strip_exponent,
+            equalize_norms=equalize_norms,
+        )
+        if strip_exponent:
+            Z = Z[0] * 10 ** Z[1]
+
+    assert Z == pytest.approx(Zex, rel=1e-3)
+
+
+@pytest.mark.parametrize("strip_exponent", [False, True])
+@pytest.mark.parametrize("equalize_norms", [False, 1.0, True])
+@pytest.mark.parametrize("inplace", [False, True])
+def test_contract_compressed_strip_exponent(
+    strip_exponent,
+    equalize_norms,
+    inplace,
+):
+    L = 6
+    tn = qtn.TN2D_rand(L, L, 2, seed=42, dist="uniform")
+    Zex = tn.contract()
+
+    if inplace:
+        tnc = tn.copy()
+        tnc.contract_(
+            optimize="greedy-compressed",
+            max_bond=4,
+            strip_exponent=strip_exponent,
+            equalize_norms=equalize_norms,
+            progbar=True,
+        )
+        Z = tnc.arrays[0] * 10**tnc.exponent
+    else:
+        Z = tn.contract(
+            optimize="greedy-compressed",
+            max_bond=4,
+            strip_exponent=strip_exponent,
+            equalize_norms=equalize_norms,
+            progbar=True,
+        )
+        if strip_exponent:
+            Z = Z[0] * 10 ** Z[1]
+
+        if equalize_norms == 1.0:
+            # undefined
+            return
+
+    assert Z == pytest.approx(Zex, rel=1e-3)
+
+
 class TestContractOpts:
     def test_contract_strategy(self):
         assert qtn.get_contract_strategy() == "greedy"